requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
bcrypt==4.0.1
pyjwt==2.8.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
